requirements.txt
openai
boto3
localstack
python-terraform
azure-identity
azure-mgmt-resource
azure-mgmt-storage
azure-storage-blob
azure-servicebus
azure-eventgrid
azure-functions
